# .migration-backup/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Students of Islamic Law Forum API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new student account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "409":
          description: Username taken
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Sign in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Send password reset code by email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordBody"
      responses:
        "200":
          description: Reset code sent (always returns ok to prevent enumeration)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForgotPasswordResponse"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password using a code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordBody"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "400":
          description: Invalid code
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Verify admin password and elevate the current session
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminLoginBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401":
          description: Invalid
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List all members (admin only)
      responses:
        "200":
          description: Members
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/User" }
        "403":
          description: Forbidden
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /admin/users/{id}/admin:
    post:
      operationId: adminSetAdmin
      tags: [admin]
      summary: Grant or revoke admin role (main admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminToggleBody" }
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "403":
          description: Forbidden
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /admin/users/{id}/active:
    post:
      operationId: adminSetActive
      tags: [admin]
      summary: Activate or deactivate a member (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminToggleBody" }
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "403":
          description: Forbidden
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /admin/posts/{id}:
    delete:
      operationId: adminDeletePost
      tags: [admin]
      summary: Delete a post (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "403":
          description: Forbidden
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  /admin/meetings/{id}:
    patch:
      operationId: adminUpdateMeeting
      tags: [admin]
      summary: Edit a live broadcast or recorded lesson (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminUpdateMeetingBody" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Meeting" }
        "403":
          description: Forbidden
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    delete:
      operationId: adminDeleteMeeting
      tags: [admin]
      summary: Delete a live broadcast or sitting (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "403":
          description: Forbidden
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
    Gender:
      type: string
      enum: [male, female]
    User:
      type: object
      required: [id, username, displayName, gender, isAdmin, isMainAdmin, isActive, createdAt]
      properties:
        id: { type: integer }
        username: { type: string }
        displayName: { type: string }
        email:
          type: ["string", "null"]
          format: email
        gender: { $ref: "#/components/schemas/Gender" }
        country:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        isAdmin: { type: boolean }
        isMainAdmin: { type: boolean }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
...
    RegisterBody:
      type: object
      required: [username, displayName, email, password, gender]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 32
          description: lowercase letters, numbers, underscores
        displayName:
          type: string
          minLength: 2
          maxLength: 60
        email:
          type: string
          format: email
          maxLength: 254
        password:
          type: string
          minLength: 6
        gender: { $ref: "#/components/schemas/Gender" }
        country:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
    LoginBody:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }
    ForgotPasswordBody:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
    ForgotPasswordResponse:
      type: object
      required: [ok, emailConfigured]
      properties:
        ok: { type: boolean }
        emailConfigured: { type: boolean }
    ResetPasswordBody:
      type: object
      required: [email, code, newPassword]
      properties:
        email: { type: string, format: email }
        code: { type: string, minLength: 4, maxLength: 12 }
        newPassword: { type: string, minLength: 6 }
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
...
    Meeting:
      type: object
      required: [id, title, scholar, kind, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        description:
          type: ["string", "null"]
        scholar: { type: string }
        kind:
          type: string
          enum: [live, recorded]
        videoUrl:
          type: ["string", "null"]
        liveUrl:
          type: ["string", "null"]
        scheduledFor:
          type: ["string", "null"]
          format: date-time
        durationMinutes:
          type: ["integer", "null"]
        coverImageUrl:
          type: ["string", "null"]
        createdByUserId:
          type: ["integer", "null"]
        createdAt: { type: string, format: date-time }
...
    AdminUpdateMeetingBody:
      type: object
      properties:
        title: { type: string, minLength: 3, maxLength: 200 }
        description:
          type: ["string", "null"]
          maxLength: 2000
        scholar: { type: string, minLength: 2, maxLength: 120 }
        kind:
          type: string
          enum: [live, recorded]
        liveUrl:
          type: ["string", "null"]
          maxLength: 500
        videoUrl:
          type: ["string", "null"]
          maxLength: 500
        scheduledFor:
          type: ["string", "null"]
          format: date-time
        durationMinutes:
          type: ["integer", "null"]
        coverImageUrl:
          type: ["string", "null"]
...
    AdminLoginBody:
      type: object
      required: [password]
      properties:
        password: { type: string }
    AdminToggleBody:
      type: object
      required: [value]
      properties:
        value: { type: boolean }
...