# .migration-backup/lib/api-zod/src/generated/types/adminUpdateMeetingBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Students of Islamic Law Forum API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminUpdateMeetingBodyKind } from "./adminUpdateMeetingBodyKind";

export interface AdminUpdateMeetingBody {
  /**
   * @minLength 3
   * @maxLength 200
   */
  title?: string;
  /**
   * @maxLength 2000
   * @nullable
   */
  description?: string | null;
  /**
   * @minLength 2
   * @maxLength 120
   */
  scholar?: string;
  kind?: AdminUpdateMeetingBodyKind;
  /**
   * @maxLength 500
   * @nullable
   */
  liveUrl?: string | null;
  /**
   * @maxLength 500
   * @nullable
   */
  videoUrl?: string | null;
  /** @nullable */
  scheduledFor?: Date | null;
  /** @nullable */
  durationMinutes?: number | null;
  /** @nullable */
  coverImageUrl?: string | null;
}