# .migration-backup/lib/api-zod/src/generated/types/loginBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Students of Islamic Law Forum API
 * OpenAPI spec version: 0.1.0
 */

export interface LoginBody {
  username: string;
  password: string;
}