# .migration-backup/lib/api-zod/src/generated/types/registerBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Students of Islamic Law Forum API
 * OpenAPI spec version: 0.1.0
 */
import type { Gender } from "./gender";

export interface RegisterBody {
  /**
   * lowercase letters, numbers, underscores
   * @minLength 3
   * @maxLength 32
   */
  username: string;
  /**
   * @minLength 2
   * @maxLength 60
   */
  displayName: string;
  /** @maxLength 254 */
  email: string;
  /** @minLength 6 */
  password: string;
  gender: Gender;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  bio?: string | null;
}